Renaming classes to use proper Camelcase and underscoring
Also renamed the Ciffilter based classes to CifBase. The class is
supposed to serve as a basis for various cif scripts that come with
the cod-tools package, one of which is cif_filter. The fact that it
was called cif_filter, despite not being aimed only at that script
was confusing. It is better to call it CifBase which will convey that
it can be used for most cif scripts directly, or be easily adapted.