Add specific calculationentry points for CifFilter and CifSelect
Even though both cif_filter and cif_select could be run with a the
CifBaseCalculation class, it is useful to provide separate calculation
classes and entry points since these scripts are used very often and
having a separate calculation class, which will show up in the provenance
graph, will be a lot clearer to the user. Now the nodes will correctly
be labeled CifFilterCalculation and CifSelectCalculation instead of
both CifBaseCalculation. This also makes querying possible.