Fix the query parameter number_species check for MPDS database
With the current implementation the only query parameter supported
is the number of species and for the MPDS database a non null value
is required. In addition, it does not support integer values, but
expects the literal string value, such as binary and terninary.
This is only defined up to quinaries