CifCleanWorkChain: make the parsing engine an optional input
The parsing engine that was used to parse the cleaned cif file for
the structure was hardcoded to be ASE, however, pymatgen is also
natively supported by the CifData class and some users may prefer
that one. Additionally, during testing, it was found that parsing
the cif with ASE could raise exceptions such as SpacegroupNotFoundError
or an AttributeError raised in method in 'tags2atoms' module.
As a solution we catch these exception and raise CifParseError.
This will cause the workchain to return the ERROR_CIF_PARSE_STRUCTURE_FAILED
error code indicating that parsing the structure failed. Note
that this will only happen if the user asked for the structure to be
parsed as that is optional