Implement the --batch-count option for cif import script
For imports of big amounts of cifs, the old implementation of storing
and adding each CifData as soon as it was parsed, is not the most
efficient. Rather, storing the parsed nodes and adding them to the
group in batches will be a lot more efficient. The downside being that
the current batch will be lost when the script is stopped or killed
unexpectedly, but that might be acceptable for large scale imports