launch_misc_cif_import: make the group option non required
If the --count-entries option is specified, the user just wants to get
the total number of entries in the query set and so a group is not needed.
Making the user pass this regardless by having it required is annoying
and unnecessary. Instead we make it optional and just check that it is
specified, *unless* the --count-entries option was passed.