CifCleanWorkChain: add useful queryable extras to structure (#36)
When generating a big database of structures, often one wants to
query for structures with a specific composition, chemical system
and or spacegroup. These properties are not stored in such a way
in the StructureData that they can be easily queried for. To solve
this, we will set these properties as extras of the primitive
structure as determined by SeeKpath in the workfunction