`launch_workflow_cif_clean`: improve query efficiency
The `launch_workflow_cif_clean` CLI needs to determine which `CifData`
nodes to launch a `CifCleanWorkChain` for, based on which ones already
have an associated workchain in the `group-workchain` while taking into
account various other filters that can be passed by the user. Here we
make the query a lot more efficient by directly including all these
filters on the database level, instead of loading all `CifData` nodes
into memory and performing the filters on the ORM level. This
drastically decreases the time needed to select the set of nodes that
should be submitted.