`launch_workflow_cif_clean`: remove the `starting-node` option
This option was added due to inefficiencies in `aiida-core`. The option
served to be able to run multiple instance of this script in parallel
without submitting the same `CifData` more than once. However, the usage
of this option was not clearly documented and was not safely implemented
risking undesirable results. Since the inefficiences that inspired the
creation of this option, have been resolved, it is better to remote it.