`launch_misc_cif_import`: make existing data query more efficient
To ensure that cif files that have already been imported are not
imported again, the list of existing source ids needs to be retrieved
that is already present in the database. This used to be done by loading
all the `CifData` nodes in the target `Group` in memory and getting the
source id through the `get_attr` method. This is now replaced by a
direct `QueryBuilder` query which is enormously more efficient.