Make `withmpi=False` the default for `get_default_options`
All of calculations defined in this plugin are simple scripts that need
to be run. They will most likely not benefit from running in MPI mode as
they are relatively cheap computations, however, there is an added risk
that the calculation will fail. Therefore it is better to have the
default be to not run with MPI.