`launch_workflow_cif_clean`: reduce duplication of input nodes
By not defining all input nodes when launching a `CifCleanWorkChain` the
workchain will create new nodes from the default values everytime. This
will bloat the database when running in high-throughput mode. Instead we
introduce a utility function that will try to find a matching existing
node with the given desired value or otherwise create one.