`CifBaseParser`: do not except if stdout cannot be parsed into `CifData` (#59)
Sometimes the stdout procuded by scripts like `cif_filter` and
`cif_select` will not always contain a correct CIF, with the consequence
that the attempted construction of a `CifData` from it in the parser
will raise an exception. An example is where if the script is run with
MPI and an error is encountered there, the process terminates after
writing the error message to stdout. The exception thrown by the parsing
into a `CifData` should be caught and transformed to appropriate exit code.