Add support for importing multinaries with MPDS in `cif_import`
A new key `multinary` has been added to the `query.classes` API
filtering key. This now allows to query and import CIF files with more
than 5 elements from the MPDS. The only problem is that we cannot
specify exactly how many elements so even if someone asks for 7
elements, all structures with more than 5 will be imported. Due to the
querybuilder check though, at least CIFs that are already imported will
not be imported again.