Add formula check to cif cleaner
Vacancies will be ignored in missing-elements formula check

Add chemical formula check to CifCleanWorkChain

A new step `check_chemical_formula` has been added to the `CifCleanWorkChain`.
The composition of the `StructuraData` created by `primitive_structure_from_cif`
in the previous step is checked against the chemical formula reported in
the `CifData` node. If it is compatible, the structure will be used as
output; otherwise an error will be raised and no structure will be
returned by the workchain. This check can be skipped by setting the
`skip_formula_check` input variable to True.