CI: move build of docs to ReadTheDocs
Using the `.readthedocs.yml` is the preferred method of configuring the
dependencies.

We also remove the dependency `click-completion` which installs `six`
which causes the build to fail. It's not quite clear why removing it
fixes it, because it is a dependency of `aiida-core` and so should be
installed regardless.

Finally, the nitpick option, which causes warnings to be turned into
errors, is removed since the following warning appears:

    cannot import name 'collections_abc' from 'six.moves' (unknown location)

It seems like this comes from an outdated `six` version but we do not
invoke `six` anywhere in the package, so it should be coming from
somewhere else.