Add extra mpds tag to `get_formula_from_cif`
The previous version of `get_formula_from_cif` would only check for the
`_pauling_file_chemical_formula`. However, there is also the
`_pauling_file_phase` tag that contains the formula in the first part of
a sequence separated by `/`.

Here we also look for this tag after parsing the `.cif`, and extract the
corresponding formula.

Finally, there is also a small refactoring of the code using the
`get_el_amt_dict` of the pymatgen `Composition` class.