Remove adding of `has_partial_occupancies` extra to `StructureData`
This extra is set in `primitive_structure_from_cif` using a function
that is a one-liner. It is then used in `check_formula`. It is better to
just run this one line of code directly in `check_formula` because in
this way it functioning properly is not dependent on the structure
having been generated by `primitive_structure_from_cif` but any
structure can be used.