Dependencies: Add support for Python 3.11 (#113)
The `pytest` dependency also had to be updated to provide support for
Python 3.11. The `isort` pre-commit version had to be upped to deal with
a build error introduced by an update in `poetry-core`. Finally, the
`pylint` dependency had to be updated since the older version depends on
a version of `astroid` which pins `wrapt<1.14`, and `wrapt~=.1.14` is
required for Python 3.11 support.